Moved the downloading and processing of artifacts to a hangfire task so we can monitor and retry if needed
